events/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.2.1
starlette==0.40.0
websockets==12.0
